python-functions/stripe_webhook/requirements.txt
stripe==5.0.0
flask==2.2.2
python-dotenv